// src/generated/billing-api/types.gen.ts
export type ErrorResponse = {
  error: {
    code: string;
    message: string;
    retryable: boolean;
    details?: {
      [key: string]: string;
    };
  };
};